Anpassung curChart

+++ KaroshiGUI/functions.h
@@ -6,4 +6,4 @@
 void getSpeedData(uint16_t speed, uint8_t* data);
 int getCanData(uint8_t* data);
 double getCanFloatData(uint8_t* data);
-double fixed_to_float(uint16_t input);
+double fixed_to_float(int32_t input);
